Refactor formatAlertText function
