+++ src/utils/common.js
@@ -17,3 +17,42 @@ export function capitalise(text) {
   const value = String(text);
   return value.charAt(0).toUpperCase() + value.slice(1);
 }
+
+export function formatAlertText(text, { formatLines = false } = {}) {
+  if (!text) return '';
+
+  const escaped = String(text)
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;');
+
+  const linkified = escaped.replace(
+    /(https?:\/\/[^\s"<>\[]+)/g,
+    (url) => {
+      const parts = url.match(/^(.*?)([\].,;:!?]*)$/);
+      const cleanUrl = parts ? parts[1] : url;
+      const suffix = parts ? parts[2] : '';
+      return `<a href="${cleanUrl}" target="_blank" rel="noopener noreferrer">${cleanUrl}</a>${suffix}`;
+    }
+  );
+
+  const normalised = linkified
+    .replace(/\\n/g, '\n')
+    .replace(/\r\n/g, '\n')
+    .replace(/\r/g, '\n');
+
+  if (!formatLines) {
+    return normalised;
+  }
+
+  return normalised
+    .split('\n')
+    .map((line) => {
+      const withBoldLabel = line.replace(
+        /^\s*([A-Za-z][^:<]{0,120}:)/,
+        '<strong>$1</strong>'
+      );
+      return `<div class="alert-line">${withBoldLabel}</div>`;
+    })
+    .join('\n');
+}
